Moved parser to use cases because it is a usecase
